requirements.txt
# Core
Django==5.1.3
djangorestframework==3.15.2
asgiref==3.8.1
sqlparse==0.4.4
tzdata==2024.2

# Documentación
drf-yasg==1.21.7

# Autenticación JWT
djangorestframework-simplejwt==5.3.1

# Filtros
django-filter==24.1

# Otros
psycopg2-binary==2.9.9  